LayoutTests/editing/deleting/ios/no-autoshift-after-deleting-character-in-word-expected.txt

Verifies that typing a word and then deleting backwards does not cause the keyboard to automatically enter shift state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isAutoShifted is false
PASS successfullyParsed is true

TEST COMPLETE
Hel
